up gif image
Top window : It’s a Francois Junod’s Mechanical Horse in Spain Project.

Under 3 windows :
I will show you an exceptional linkage here. to α2α3, I named “Divine
joint”.
it is a joint which include plenty of unpredictability and fun. for
examples, organic curves of the approximate airfoil of birds fuselage
will be obtained by the formula who added this joint.