Details of demo 4 development by gif animation (Improvement)
Top :
François Junod’s Mechanical Horse in Spain project.
There is many amazing creativity.
It is also like the Cambrian Explosion of mechanical elements.

Bottom threes :
I show an exceptional linkage of the formula of voluntary movement here.
to α2α3, I named “Divine joint”.
it is a joint which include plenty of unpredictability and fun.
for examples, organic curves of the approximate airfoil of birds
fuselage will be obtained by the formula who added this joint.
Inverse Kinematcs is a necessary technique to analysis for increased
uncertainty. in here, I used CCD algorithm for solving it.